Changed next_sector to throw on bad input
The returned position can be 1 sector beyond the end of the disk, but not
the input. If callers attempt to use the invalid output they'll fail
elsewhere. Added unit tests to check invalid inputs.

Also changed disk writing to detect this failure instead of the sector write
being out of range. Data is now written before the directory entry too, so
full disks don't contain a truncated file.