Added support for DISCiPLE .img disk image files
DISCiPLE .img files are the same size as .mgt but use a different track order.
All tracks on head 0 are stored first, followed by all tracks on head 1.

If there's a file on the disk we can usually check the sector chain to detect
.img format. If the file is only a single sector there's no chain, but we can
check the header in the directory entry against the first 9 bytes of the file.