Changed to use sector bitmap for more operations
File reading and writing now uses only the sector bitmap, which is more robust
on damaged disks that break the sector chain. Chain sectors are always in the
same order as the bitmap -- MasterDOS takes special care to add after the last
position when adding to OUT/OPENTYPE files, in case files were deleted.

Sectors are now allocated using the disk sector bitmap as part of
File.to_dir(). This handles skipping the extra directory tracks and the special
case for the boot sector, and could support adding files to existing disks in
the future.

Replaced start_track and start_sector with first_sector, which is now a
property returning a tuple. The value is calculated from the first bit in the
file bitmap. Any start track/sector read from an existing disk is stored in
File._first_sector, which is not used but available for inspection.

The file header data is now generated as part of to_dir, for SAM types that
need it (ZX already sets them). The File.header is extracted from the directory
entry, and will be written before the file data where appropriate.