Added sector reservation for Disk.to_image()
The reserved_map parameter is a sector bitmap covering all 1600 sectors of a
standard disk. Sectors can be reserved in both the directory and data areas,
which will be seen as in-use when placing files.

Reserved directory sectors must not be completely empty or many tools will
treat them as the end of the directory. Setting the second character of each
256-byte entry to anything other than 0x00 should be enough.