Added support for fixed SPECIAL files
SPECIAL type files tend to be used for situations where their disk position is
important. They don't use a sector chain so the full sector can be used to hold
data.

SPECIAL files now retain their existing sector map, so they'll be written to
the same position on the disk. Files with an empty sector map will be allocated
space, so it can be cleared to force reallocation (at your own risk!).